+++ src/commands/commit.ts
@@ -48,27 +48,23 @@ export class Commit extends Command {
         }
       }
 
-      try {
-        const message = await inputCommitMessage(
-          repository.inputBox.value,
-          true,
-          paths
-        );
+      const message = await inputCommitMessage(
+        repository.inputBox.value,
+        true,
+        paths
+      );
 
-        if (message === undefined) {
-          return;
-        }
+      if (message === undefined) {
+        return;
+      }
 
-        repository.inputBox.value = message;
+      repository.inputBox.value = message;
 
+      await this.handleRepositoryOperation(async () => {
         const result = await repository.commitFiles(message, paths);
         window.showInformationMessage(result);
         repository.inputBox.value = "";
-      } catch (error) {
-        console.error(error);
-        const svnError = error as ISvnErrorData;
-        window.showErrorMessage(svnError.stderrFormated || String(error));
-      }
+      }, "Unable to commit");
     });
   }
 }