+++ src/commands/patch.ts
@@ -7,11 +7,8 @@ export class Patch extends Command {
   }
 
   public async execute(...resourceStates: SourceControlResourceState[]) {
-    const selection = await this.getResourceStates(resourceStates);
-
-    if (selection.length === 0) {
-      return;
-    }
+    const selection = await this.getResourceStatesOrExit(resourceStates);
+    if (!selection) return;
 
     const uris = selection.map(resource => resource.resourceUri);