+++ src/commands/resolveAll.ts
@@ -26,16 +26,13 @@ export class ResolveAll extends Command {
         return;
       }
 
-      try {
+      await this.handleRepositoryOperation(async () => {
         const response = await repository.resolve(
           [conflict.resourceUri.path],
           choice.label
         );
         window.showInformationMessage(response);
-      } catch (error) {
-        const svnError = error as ISvnErrorData;
-        window.showErrorMessage(svnError.stderr || String(error));
-      }
+      }, "Unable to resolve conflict");
     }
   }
 }